Absorb API contract into types
