+++ crates/perfgate-cli/Cargo.toml
@@ -20,7 +20,6 @@ path = "src/main.rs"
 [dependencies]
 perfgate-types.workspace = true
 perfgate-client.workspace = true
-perfgate-api.workspace = true
 perfgate-app.workspace = true
 perfgate-adapters.workspace = true
 perfgate-domain.workspace = true
