+++ crates/perfgate-client/Cargo.toml
@@ -16,7 +16,6 @@ categories = ["web-programming::http-client", "development-tools"]
 [dependencies]
 anyhow.workspace = true
 perfgate-types.workspace = true
-perfgate-api.workspace = true
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-webpki-roots"] }
 tokio = { version = "1", features = ["time", "fs", "io-util"] }
 serde = { version = "1", features = ["derive"] }
