+++ crates/perfgate-client/src/types.rs
@@ -1,5 +1,5 @@
 //! Request and response types for the baseline service API.
 //!
-//! These types are re-exported from `perfgate-api` to maintain the API contract.
+//! These types are re-exported from `perfgate-types` to maintain the API contract.
 
-pub use perfgate_api::*;
+pub use perfgate_types::baseline_service::*;
