+++ crates/perfgate-server/Cargo.toml
@@ -45,7 +45,6 @@ sqlx = { version = "0.8.3", default-features = false, features = ["postgres", "r
 
 # Validation
 perfgate-types.workspace = true
-perfgate-api = { workspace = true, features = ["server"] }
 perfgate-domain.workspace = true
 
 # Error handling
@@ -55,6 +54,7 @@ anyhow = { workspace = true }
 # CLI
 clap = { workspace = true }
 regex = { workspace = true }
+toml.workspace = true
 
 # Metrics
 metrics = "0.24.3"
