+++ crates/perfgate-server/src/credential_source.rs
