+++ crates/perfgate-types/Cargo.toml
@@ -24,6 +24,8 @@ schemars.workspace = true
 thiserror.workspace = true
 toml.workspace = true
 toml_edit = "0.23"
+chrono = { version = "0.4.39", features = ["serde"] }
+uuid.workspace = true
 arbitrary = { version = "1", features = ["derive"], optional = true }
 
 [dev-dependencies]
