+++ docs/MIGRATION_0_16.md
@@ -74,9 +74,9 @@ Use this table to find the new import path for any old crate:
 | `perfgate-summary` | `perfgate::presentation::render::summary` | `use perfgate::presentation::render::summary::*;` |
 | `perfgate-error` | `perfgate_types::error` | `use perfgate_types::error::*;` |
 | `perfgate-validation` | `perfgate_types::validation` | `use perfgate_types::validation::*;` |
-| `perfgate-auth` | `perfgate_api::auth` now; final owner TBD | `use perfgate_api::auth::*;` |
+| `perfgate-auth` | `perfgate_types::baseline_service::auth` | `use perfgate_types::baseline_service::auth::*;` |
 | `perfgate-config` | `perfgate_types::config` and `perfgate_client::ResolvedServerConfig` | `use perfgate_types::config::*; use perfgate_client::ResolvedServerConfig;` |
-| `perfgate-api` | `perfgate_types::baseline_service` | `use perfgate_types::baseline_service::*;` |
+| `perfgate-api` | `perfgate_types::baseline_service`; server credential loading in `perfgate_server::CredentialSource` | `use perfgate_types::baseline_service::*;` |
 | `perfgate-fake` | private workspace crate | No public replacement yet; keep local test helpers in your own crate. |
 | `perfgate-adapters` | `perfgate::runtime` | `use perfgate::runtime::*;` |
 | `perfgate-app` | `perfgate::app` | `use perfgate::app::*;` |
