+++ docs/REFACTORING_0_16.md
@@ -4,7 +4,7 @@
 
 This document describes the planned refactoring of perfgate's public crate surface. The goal is to collapse the current 26+ microcrates into a cleaner public facade while preserving the strong internal separation of concerns that makes the architecture maintainable.
 
-**Current state**: Public surface is still too broad. PR #223 already absorbed `perfgate-validation`, `perfgate-auth`, `perfgate-summary`, and `perfgate-stats`, and moved the paired implementation into `perfgate-domain` behind a workspace-only `perfgate-paired` compatibility wrapper. The error contract now lives in `perfgate_types::error` behind a workspace-only `perfgate-error` compatibility wrapper, profiling now lives under `perfgate::runtime::profile`, external benchmark ingestion now lives under `perfgate::integrations::ingest`, deterministic fingerprinting now lives in `perfgate_types::fingerprint` with a `perfgate::core::fingerprint` facade path, host mismatch detection now lives in `perfgate_domain::host` with a `perfgate::domain::host` facade path, budget evaluation now lives in `perfgate_domain::budget` with a `perfgate::core::budget` facade path, scaling analysis now lives in `perfgate_domain::scaling` with a `perfgate::domain::scaling` facade path, rendering now has the facade path `perfgate::presentation::render` behind a workspace-only `perfgate-render` compatibility wrapper, export now has the facade path `perfgate::presentation::export` behind a workspace-only `perfgate-export` compatibility wrapper, sensor report generation now has the facade path `perfgate::presentation::sensor` behind a workspace-only `perfgate-sensor` compatibility wrapper, and GitHub PR-comment integration now has the feature-gated facade path `perfgate::integrations::github` behind a workspace-only `perfgate-github` compatibility wrapper. Many remaining internal seams are still publishable crates (`perfgate-api`, `perfgate-domain`, etc.), which makes the package ecosystem confusing for users and couples the public API tightly to internal refactoring decisions.
+**Current state**: Public surface is still too broad. PR #223 already absorbed `perfgate-validation`, `perfgate-auth`, `perfgate-summary`, and `perfgate-stats`, and moved the paired implementation into `perfgate-domain` behind a workspace-only `perfgate-paired` compatibility wrapper. The error contract now lives in `perfgate_types::error` behind a workspace-only `perfgate-error` compatibility wrapper, profiling now lives under `perfgate::runtime::profile`, external benchmark ingestion now lives under `perfgate::integrations::ingest`, deterministic fingerprinting now lives in `perfgate_types::fingerprint` with a `perfgate::core::fingerprint` facade path, host mismatch detection now lives in `perfgate_domain::host` with a `perfgate::domain::host` facade path, budget evaluation now lives in `perfgate_domain::budget` with a `perfgate::core::budget` facade path, scaling analysis now lives in `perfgate_domain::scaling` with a `perfgate::domain::scaling` facade path, rendering now has the facade path `perfgate::presentation::render` behind a workspace-only `perfgate-render` compatibility wrapper, export now has the facade path `perfgate::presentation::export` behind a workspace-only `perfgate-export` compatibility wrapper, sensor report generation now has the facade path `perfgate::presentation::sensor` behind a workspace-only `perfgate-sensor` compatibility wrapper, GitHub PR-comment integration now has the feature-gated facade path `perfgate::integrations::github` behind a workspace-only `perfgate-github` compatibility wrapper, and baseline-service API contracts now live in `perfgate_types::baseline_service` behind a workspace-only `perfgate-api` compatibility wrapper. Many remaining internal seams are still publishable crates (`perfgate-domain`, `perfgate-app`, etc.), which makes the package ecosystem confusing for users and couples the public API tightly to internal refactoring decisions.
 
 **Target state**: Five public crates with strongly organized internal modules. Users depend on `perfgate`, `perfgate-types`, `perfgate-cli`, `perfgate-client`, and `perfgate-server` only. The SRP boundaries remain enforced but move from crate level to module level.
 
@@ -38,7 +38,7 @@ These are contract-adjacent and belong with the public receipt/config model:
 | `perfgate-config` | `perfgate_types::config` + `perfgate_client::ResolvedServerConfig` | Config model and file helpers are contract-adjacent; client construction belongs to the client crate |
 | `perfgate-api` (shared DTOs) | `perfgate_types::baseline_service` | Wire format for baseline service |
 
-**Current state**: `perfgate-validation` and `perfgate-config` have been deleted, and `perfgate-error` now re-exports `perfgate_types::error` as a workspace-only compatibility wrapper. Config file helpers now live in `perfgate_types::config`, resolved baseline-server client construction lives in `perfgate_client`, and shared baseline-service DTOs are the remaining contract-adjacent seam.
+**Current state**: `perfgate-validation` and `perfgate-config` have been deleted, and `perfgate-error` now re-exports `perfgate_types::error` as a workspace-only compatibility wrapper. Config file helpers now live in `perfgate_types::config`, resolved baseline-server client construction lives in `perfgate_client`, shared baseline-service DTOs now live in `perfgate_types::baseline_service`, and server credential-source loading now lives in `perfgate_server::CredentialSource`.
 
 **Why first**: The types crate must be standalone and self-describing. Absorbing these dependencies first unblocks all downstream refactoring.
 
@@ -294,7 +294,7 @@ Move into `perfgate-types`:
 - `perfgate-error` -> `perfgate_types::error`
 - `perfgate-validation` -> `perfgate_types::validation`
 - `perfgate-config` -> `perfgate_types::config` and `perfgate_client::ResolvedServerConfig`
-- `perfgate-api` shared DTOs -> `perfgate_types::baseline_service`
+- `perfgate-api` shared DTOs -> already landed in `perfgate_types::baseline_service`; server credential-source loading moved to `perfgate_server::CredentialSource`
 
 Update all downstream crate imports. Verify `perfgate-types` remains self-contained.
